pallet_ah_ops::pallet::dispatchables

Function migrate_parachain_sovereign_derived_acc

pub fn migrate_parachain_sovereign_derived_acc<T: Config>(
    from: T::AccountId,
    to: T::AccountId,
    derivation: (T::AccountId, DerivationIndex),
)

Try to migrate a parachain sovereign child account to its respective sibling.

Takes the old and new account and migrates it only if they are as expected. An event of SovereignMigrated will be emitted if the account was migrated successfully.

Callable by any signed origin.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::migrate_parachain_sovereign_derived_acc.